requirements.txt
nemo_toolkit[asr]==2.4.0
torchaudio==2.4.0
fastapi==0.115.12
uvicorn[standard]==0.30.6
uvloop==0.19.0
websockets==12.0
soundfile==0.12.1
pydub==0.25.1
huggingface_hub[hf-xet]==0.31.2
hf_transfer==0.1.9
numpy<2
